añadido readme y renombrado cache_time.py a app_time.py

+++ lib/problema2/jefe_mesa_problema2.h
@@ -28,7 +28,7 @@ extern int cartas[10];
 extern int n_disponibles;
 extern int jugando;
 extern int cartas_disponibles;
-extern long sem_jugadores_value;
+extern int sem_jugadores_value;
 
 // mutex
